ai_takeaki_for_github/requirements.txt
streamlit
langchain
chromadb
openai
python-dotenv
docx2txt
